Migrate test files to component path
