+++ components/channel_header/components/header_icon_wrapper.test.js
