+++ components/channel_header_dropdown/menu_items/convert_channel.test.js
@@ -4,17 +4,20 @@
 import React from 'react';
 import {shallow} from 'enzyme';
 
-import {NotificationPreferences} from 'components/channel_header_dropdown/menu_items';
+import ConvertChannel from './convert_channel';
 
-describe('components/ChannelHeaderDropdown/MenuItem.NotificationPreferences', () => {
+describe('components/ChannelHeaderDropdown/MenuItem.ConvertChannel', () => {
     it('should match snapshot', () => {
         const wrapper = shallow(
-            <NotificationPreferences
-                user={{}}
-                channel={{}}
-                membership={{}}
+            <ConvertChannel
+                channel={{
+                    id: 'channel_id',
+                    team_id: 'team_id',
+                    display_name: 'Test',
+                }}
             />
         );
+
         expect(wrapper).toMatchSnapshot();
     });
 });
