+++ components/channel_header_dropdown/menu_items/view_members.test.js
@@ -5,8 +5,11 @@ import React from 'react';
 import {shallow} from 'enzyme';
 import {Permissions} from 'mattermost-redux/constants';
 
-import {ManageMembers} from 'components/channel_header_dropdown/menu_items';
 import {Constants} from 'utils/constants';
+
+import ChannelPermissionGate from 'components/permissions_gates/channel_permission_gate';
+
+import ViewMembers from './view_members';
 
 describe('components/ChannelHeaderDropdown/MenuItem.ViewMembers', () => {
     const baseProps = {
@@ -19,24 +22,24 @@ describe('components/ChannelHeaderDropdown/MenuItem.ViewMembers', () => {
     };
 
     it('should match snapshot', () => {
-        const wrapper = shallow(<ManageMembers {...baseProps}/>);
+        const wrapper = shallow(<ViewMembers {...baseProps}/>);
 
         expect(wrapper).toMatchSnapshot();
     });
 
-    it('should be hidden if the channel is default channel', () => {
+    it('should appears without permission check if the channel is default channel', () => {
         const props = {
             ...baseProps,
             isDefault: true,
         };
-        const wrapper = shallow(<ManageMembers {...props}/>);
+        const wrapper = shallow(<ViewMembers {...props}/>);
 
-        expect(wrapper.isEmptyRender()).toBeTruthy();
+        expect(wrapper.is(ChannelPermissionGate)).toBeFalsy();
     });
 
     it('should right permission level for channel type', () => {
         const props = baseProps;
-        const makeWrapper = () => shallow(<ManageMembers {...props}/>);
+        const makeWrapper = () => shallow(<ViewMembers {...props}/>);
 
         props.channel.type = Constants.OPEN_CHANNEL;
         expect(makeWrapper().prop('permissions')[0]).toBe(Permissions.MANAGE_PUBLIC_CHANNEL_MEMBERS);
